OBPIH-6013 Make Tabs component reusable
